changed polygon side to u32 + set max polygon side to 1000

+++ graphene/src/operation.rs
@@ -95,7 +95,7 @@ pub enum Operation {
 		path: Vec<LayerId>,
 		insert_index: isize,
 		transform: [f64; 6],
-		sides: u8,
+		sides: u32,
 		style: style::PathStyle,
 	},
 	AddOverlayShape {
@@ -215,8 +215,8 @@ impl Operation {
 	/// # Safety
 	/// This function reads from uninitialized memory!!!
 	/// Only use if you know what you are doing
-	unsafe fn as_slice(&self) -> &[u8] {
-		core::slice::from_raw_parts(self as *const Operation as *const u8, std::mem::size_of::<Operation>())
+	unsafe fn as_slice(&self) -> &[u32] {
+		core::slice::from_raw_parts(self as *const Operation as *const u32, std::mem::size_of::<Operation>())
 	}
 	/// Returns a pseudo hash that should uniquely identify the operation.
 	/// This is needed because `Hash` is not implemented for f64s
